Updated f=0 amplitude compensation
Until now if the frequency of a given voice was f=0Hz then I've set both
the DCO freq. and the amplitude compensation to zero. This however
resulted a negative DC offset on the output of the DCO. Instead if I set
_only_ the amplitude to zero (and use another voice's frequency to
reset the given DCO) then the DC output is closer to 0 which helps
minimising the pop when voices are turned on.