DCO schematic update
Changed the DCO design from using a PNP transistor for the switch,
positive charge and an inverting amp with bias for the saw output. This
allows to bias the sawtooth output to oscillate around 0V which makes
possible to set precise pulse width values for all voices regardless of
potential component differences. Also, this eventually needs less
component altogether.