Added solo submode
ATM the plan is to have only a single VCA and ADSR which brings up the
question of how to handle paraphonic mode. The question is what to do
with the notes which are released on the keyboard and how that would
work with the envelope.

The core idea is that voice #1 will control the envelope, ie. when a new
note is played on voice #1 that'll trigger a new gate out. When more
than one note is played, voices will turn on for each one of them.
However, once a non-first note is released while the first one is held
we can do two things:

1. Set the amplitude of the released (non-first) note to zero. The pro
of this is that it'll act as expected (released note should _not_ play)
however it'll shut off immediately, ie. it'll not have any envelope.
2. Do nothing, ie. let the voice keep playing. The pros and cons are the
opposite of option #1: once the first voice is released all playing
voices will be processed by the envelope, but it has the kinda awkward
sideeffect of playing a note even though it's released on the keyboard.

From the usage point of view, none is better than the other: option #1
is better for soloing, option #2 is better for chords.

This commit adds the possibility to set between these two submodes.